Add argparse demo with data and example output
This commit creates a very basic (probably buggy) argparse demo. The
included data files are scipy.stats models and there are two versions of
a script that do more-or-less the same thing: one employs argparse, the
other is written in the style of someone who hasn't yet learned the
magic of argparsing.